requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
python-multipart==0.0.20
pymupdf==1.25.5
pdf2image==1.17.0
pdf2docx==0.5.8
python-pptx==1.0.2
fpdf==1.7.2
python-dotenv==1.0.0
